+++ packages/dothtml/CHANGELOG.md
@@ -1,5 +1,13 @@
 # dothtml
 
+## 6.0.6
+
+### Patch Changes
+
+- Fix styling typings + bug fix for empty().
+- Updated dependencies
+  - dothtml-interfaces@6.0.6
+
 ## 6.0.5
 
 ### Patch Changes