+++ packages/interfaces/CHANGELOG.md
@@ -1,5 +1,11 @@
 # dothtml-interfaces
 
+## 6.0.6
+
+### Patch Changes
+
+- Fix styling typings + bug fix for empty().
+
 ## 6.0.5
 
 ### Patch Changes